Complete Lab1
