+++ Lab1/shell.h
@@ -1,0 +1,2 @@
+int strcmp(char *s1, char *s2);
+void shell(char *cmd);
